intermittent error bugfix (#435)
* fix compiler warnings

* clang tidy warnings fixed

* * try to avoid implicit constructor and make sure offset is always initialized
* check Index lengths

* * make sure that parents length is the same as index length
* add more tests

* * declare constructors explicit and classes final
* define explicitly derived Form copy constructors - pybind11 uses them
* define default values
* remove asserts

* remove default member initializers, use member initializer list of the constructor

* actual bugfix

* fix documentation issue for a newly introduced parameter

* remove guesswork

* add more tests

* Cast integers to specialized types unless the RapidJSON API specifically calls for an 'int'.

+++ src/libawkward/array/NumpyArray.cpp
@@ -544,10 +544,10 @@ namespace awkward {
           out << (kernel::NumpyArray_getitem_at0(ptr_lib, ptr2) != 0 ? "true" : "false");
         }
         else if (dtype == util::dtype::int8) {
-          out << (int)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
+          out << (int64_t)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
         }
         else if (dtype == util::dtype::uint8) {
-          out << (unsigned int)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
+          out << (uint64_t)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
         }
         else {
           out << kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
@@ -565,10 +565,10 @@ namespace awkward {
           out << (kernel::NumpyArray_getitem_at0(ptr_lib, ptr2) != 0 ? "true" : "false");
         }
         else if (dtype == util::dtype::int8) {
-          out << (int)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
+          out << (int64_t)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
         }
         else if (dtype == util::dtype::uint8) {
-          out << (unsigned int)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
+          out << (uint64_t)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
         }
         else {
           out << kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
@@ -585,10 +585,10 @@ namespace awkward {
           out << (kernel::NumpyArray_getitem_at0(ptr_lib, ptr2) != 0 ? "true" : "false");
         }
         else if (dtype == util::dtype::int8) {
-          out << (int)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
+          out << (int64_t)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
         }
         else if (dtype == util::dtype::uint8) {
-          out << (unsigned int)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
+          out << (uint64_t)kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
         }
         else {
           out << kernel::NumpyArray_getitem_at0(ptr_lib, ptr2);
@@ -748,7 +748,7 @@ namespace awkward {
           unsigned char* ptr2 = reinterpret_cast<unsigned char*>(
               reinterpret_cast<ssize_t>(data()) + (ssize_t)i);
           out << std::hex << std::setw(2) << std::setfill('0')
-              << int((unsigned char)kernel::NumpyArray_getitem_at0(ptr_lib_, ptr2));
+              << (int64_t)((unsigned char)kernel::NumpyArray_getitem_at0(ptr_lib_, ptr2));
         }
       }
       else {
@@ -759,7 +759,7 @@ namespace awkward {
           unsigned char* ptr2 = reinterpret_cast<unsigned char*>(
               reinterpret_cast<ssize_t>(data()) + (ssize_t)i);
           out << std::hex << std::setw(2) << std::setfill('0')
-              << int((unsigned char)kernel::NumpyArray_getitem_at0(ptr_lib_, ptr2));
+              << (int64_t)((unsigned char)kernel::NumpyArray_getitem_at0(ptr_lib_, ptr2));
         }
         out << " ... ";
         for (ssize_t i = len - 16;  i < len;  i++) {
@@ -769,7 +769,7 @@ namespace awkward {
           unsigned char* ptr2 = reinterpret_cast<unsigned char*>(
               reinterpret_cast<ssize_t>(data()) + (ssize_t)i);
           out << std::hex << std::setw(2) << std::setfill('0')
-              << int((unsigned char)kernel::NumpyArray_getitem_at0(ptr_lib_, ptr2));
+              << (int64_t)((unsigned char)kernel::NumpyArray_getitem_at0(ptr_lib_, ptr2));
         }
       }
     }
